requirements.txt
openai==1.72.0
python-dotenv==1.1.0
rich==14.0.0
scikit-learn==1.6.1
numpy==2.2.4
tiktoken==0.9.0
google-genai==1.10.0
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv>=0.19.0
pydantic>=2.0.0
websockets>=10.0
redis>=4.0.0
rich>=10.0.0
aiofiles>=0.8.0
python-multipart>=0.0.5
docker>=5.0.0 
